requirements.txt
# Core packages
Django>=5.1,<5.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
django-filter>=21.1

# Testing dependencies
requests>=2.28.0
tabulate>=0.9.0

# Optional dependencies
Pillow>=9
